for issue 94 - Tensorflow loads when loading TorchClient
