+++ grid/lib/utils.py
@@ -1,7 +1,6 @@
 from filelock import FileLock
 from grid import ipfsapi
 from pathlib import Path
-import keras
 import os
 import json
 import time
@@ -52,16 +51,6 @@ def save_adapter(addr):
     adapter_bin = get_ipfs_api().cat(addr)
     ensure_exists(f'{Path.home()}/grid/adapters/adapter.py', adapter_bin)
 
-
-def keras2ipfs(model):
-    return get_ipfs_api().add_bytes(serialize_keras_model(model))
-
-
-def ipfs2keras(model_addr):
-    model_bin = get_ipfs_api().cat(model_addr)
-    return deserialize_keras_model(model_bin)
-
-
 def serialize_numpy(tensor):
     # nested lists with same data, indices
     return json.dumps(tensor.tolist())
@@ -69,78 +58,6 @@ def serialize_numpy(tensor):
 
 def deserialize_numpy(json_array):
     return np.array(json.loads(json_array)).astype('float')
-
-
-def serialize_torch_model(model, **kwargs):
-    """
-    kwargs are the arguments needed to instantiate the model
-    """
-    state = {'state_dict': model.state_dict(), 'kwargs': kwargs}
-    torch.save(state, 'temp_model.pth.tar')
-    with open('temp_model.pth.tar', 'rb') as f:
-        model_bin = f.read()
-    return model_bin
-
-
-def deserialize_torch_model(model_bin, model_class, **kwargs):
-    """
-    model_class is needed since PyTorch uses pickle for serialization
-        see https://discuss.pytorch.org/t/loading-pytorch-model-without-a-code/12469/2 for details
-    kwargs are the arguments needed to instantiate the model from model_class
-    """
-    with open('temp_model2.pth.tar', 'wb') as g:
-        g.write(model_bin)
-    state = torch.load()
-    model = model_class(**state['kwargs'])
-    model.load_state_dict(state['state_dict'])
-    return model
-
-
-def serialize_keras_model(model):
-    lock = FileLock('temp_model.h5.lock')
-    with lock:
-        model.save('temp_model.h5')
-        with open('temp_model.h5', 'rb') as f:
-            model_bin = f.read()
-            f.close()
-        return model_bin
-
-
-def deserialize_keras_model(model_bin):
-    lock = FileLock('temp_model2.h5.lock')
-    with lock:
-        with open('temp_model2.h5', 'wb') as g:
-            g.write(model_bin)
-            g.close()
-        model = keras.models.load_model('temp_model2.h5')
-        return model
-
-
-def save_best_model_for_task(task, model):
-    ensure_exists(f'{Path.home()}/.openmined/models.json', {})
-    with open(f"{Path.home()}/.openmined/models.json", "r") as model_file:
-        models = json.loads(model_file.read())
-
-    models[task] = keras2ipfs(model)
-
-    with open(f"{Path.home()}/.openmined/models.json", "w") as model_file:
-        json.dump(models, model_file)
-
-
-def best_model_for_task(task, return_model=False):
-    if not os.path.exists(f'{Path.home()}/.openmined/models.json'):
-        return None
-
-    with open(f'{Path.home()}/.openmined/models.json', 'r') as model_file:
-        models = json.loads(model_file.read())
-        if task in models.keys():
-            if return_model:
-                return ipfs2keras(models[task])
-            else:
-                return models[task]
-
-    return None
-
 
 def load_task(name):
     if not os.path.exists(f'{Path.home()}/.openmined/tasks.json'):
@@ -152,7 +69,6 @@ def load_task(name):
     for task in tasks:
         if task['name'] == name:
             return task
-
 
 def store_task(name, address):
     ensure_exists(f'{Path.home()}/.openmined/tasks.json', [])
@@ -171,10 +87,8 @@ def store_task(name, address):
         with open(f"{Path.home()}/.openmined/tasks.json", "w") as task_file:
             json.dump(tasks, task_file)
 
-
 def store_whoami(info):
     ensure_exists(f'{Path.home()}/.openmined/whoami.json', info)
-
 
 def load_whoami():
     if not os.path.exists(f'{Path.home()}/.openmined/whoami.json'):
@@ -182,7 +96,6 @@ def load_whoami():
 
     with open(f'{Path.home()}/.openmined/whoami.json', 'r') as cb:
         return json.loads(cb.read())
-
 
 def ensure_exists(path, default_contents=None):
     """
